+++ cmp/form/FormModel.ts
@@ -4,14 +4,22 @@
  *
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
-import {HoistModel, managed, PlainObject} from '@xh/hoist/core';
+import {
+    HoistModel,
+    managed,
+    PersistableState,
+    PersistenceProvider,
+    PersistOptions,
+    PlainObject
+} from '@xh/hoist/core';
 import {ValidationState} from '@xh/hoist/data';
 import {action, bindable, computed, makeObservable, observable} from '@xh/hoist/mobx';
 import {throwIf} from '@xh/hoist/utils/js';
-import {flatMap, forEach, forOwn, map, mapValues, pickBy, some, values} from 'lodash';
+import {flatMap, forEach, forOwn, isString, map, mapValues, pickBy, some, values} from 'lodash';
 import {BaseFieldConfig, BaseFieldModel} from './field/BaseFieldModel';
 import {FieldModel} from './field/FieldModel';
 import {SubformsFieldConfig, SubformsFieldModel} from './field/SubformsFieldModel';
+import {LocalDate} from '@xh/hoist/utils/datetime';
 
 export interface FormConfig {
     /**
@@ -22,11 +30,21 @@ export interface FormConfig {
     /** Map of initial values for fields in this model. */
     initialValues?: PlainObject;
 
+    /** Options governing persistence of the form state. */
+    persistWith?: FormPersistOptions;
+
     disabled?: boolean;
     readonly?: boolean;
 
     /** @internal */
     xhImpl?: boolean;
+}
+
+export interface FormPersistOptions extends PersistOptions {
+    /** If persisting only a subset of all fields, provide an array of field names. */
+    fieldsToInclude?: string[];
+    /** If excluding a subset of all fields, provide an array of field names. */
+    fieldsToExclude?: string[];
 }
 
 /**
@@ -87,6 +105,7 @@ export class FormModel extends HoistModel {
         fields = [],
         initialValues = {},
         disabled = false,
+        persistWith = null,
         readonly = false,
         xhImpl = false
     }: FormConfig = {}) {
@@ -97,6 +116,7 @@ export class FormModel extends HoistModel {
         this.disabled = disabled;
         this.readonly = readonly;
         const models = {};
+
         fields.forEach((f: any) => {
             const model =
                     f instanceof BaseFieldModel
@@ -111,6 +131,7 @@ export class FormModel extends HoistModel {
         this.fields = models;
 
         this.init(initialValues);
+        if (persistWith) this.initPersist(persistWith);
 
         // Set the owning formModel *last* after all fields in place with data.
         // This (currently) kicks off the validation and other reactivity.
@@ -272,6 +293,67 @@ export class FormModel extends HoistModel {
                     return undefined;
                 }
             }
+        );
+    }
+
+    private initPersist({
+        fieldsToInclude = null,
+        fieldsToExclude = null,
+        path = 'formValues',
+        ...rootPersistWith
+    }: FormPersistOptions) {
+        const fieldNameMap = this.createFormFieldsToPersistListing(
+            fieldsToInclude,
+            fieldsToExclude
+        );
+
+        PersistenceProvider.create({
+            persistOptions: {
+                path,
+                ...rootPersistWith
+            },
+            target: {
+                getPersistableState: () => {
+                    return new PersistableState(
+                        fieldNameMap.reduce(
+                            (acc, name) => ({...acc, [name]: this.fields[name].value}),
+                            {}
+                        )
+                    );
+                },
+                setPersistableState: ({value: formValues}) => {
+                    // There is no metadata on a field to denote it is a date.
+                    // Use a regex matcher to tests for dates and format matches accurately.
+                    this.setValues(
+                        fieldNameMap.reduce((acc, name) => {
+                            if (
+                                isString(formValues[name]) &&
+                                formValues[name].match(/^\d{4}-\d{2}-\d{2}$/)
+                            ) {
+                                return {...acc, [name]: LocalDate.from(formValues[name])};
+                            }
+                            return {...acc, [name]: formValues[name]};
+                        }, {} as PlainObject)
+                    );
+                }
+            },
+            owner: this
+        });
+    }
+    /**
+     * Helper to build a list of fields to persist on a form.
+     */
+    private createFormFieldsToPersistListing(
+        fieldsToInclude: string[] | null,
+        fieldsToExclude: string[] | null
+    ) {
+        if (!fieldsToInclude && !fieldsToExclude) return Object.keys(this.fields);
+        if (fieldsToInclude && !fieldsToExclude) return fieldsToInclude;
+        if (!fieldsToInclude && fieldsToExclude)
+            return Object.keys(this.fields).filter(name => !fieldsToExclude.includes(name));
+
+        return Object.keys(this.fields).filter(
+            name => !fieldsToExclude.includes(name) && fieldsToInclude.includes(name)
         );
     }
 }