+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v73.0.0-SNAPSHOT - unreleased
 
 * Support for reporting Client Version in Admin WebSockets tab.
+* Update the `FormModel` with configuration options to support persisting the form state - either the entire form
+or a provided subset of fields.
 
 ## v72.2.0 - 2025-03-13